requirements.txt
Django==3.2.13
django-redis==5.2.0
djangorestframework==3.13.1
Pillow==9.1.1
python-dotenv==0.20.0
flake8==4.0.1
flake8-isort==4.1.1
black==22.3.0
twilio==7.9.1
django-phonenumbers==1.0.1
django-phonenumber-field==6.1.0
Babel==2.10.3
tzdata==2022.1
channels==3.0.4
channels-redis==3.4.0
django-timezone-field==5.0
gunicorn==20.1.0
coverage==6.3.2
psycopg2-binary==2.9.3
pytest-django==4.5.2
pytest-factoryboy==2.5.0